break down consensus.go
